pyproject.toml
[project]
name = "uv-langraph-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grandalf>=0.8",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-openai>=1.1.6",
    "langchain-tavily>=0.2.15",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = []
